Added indirect block pointers

+++ include/kernel/fs/vfs.h
@@ -4,6 +4,6 @@
 #include "../include/drivers/ata.h"
 
 void vfs_probe(ata_dev_t *dev);
-uint8_t vfs_read(char* path, uint16_t *buf);
+uint8_t vfs_read(char* path, uint8_t *buf);
 
 #endif
