+++ kernel/fs/vfs.c
@@ -71,7 +71,7 @@ void vfs_probe(ata_dev_t *dev)
 	free(gpt_table_buf);
 }
 
-uint8_t vfs_read(char* path, uint16_t *buf)
+uint8_t vfs_read(char* path, uint8_t *buf)
 {
 	if (!root_dev->ready) return 0;
